Arguments that are shared by multiple wrapper functions

Arguments that are shared by multiple wrapper functions

Arguments

data

Data frame to use for plot.

x_str

Character string. Name of variable in data to plot on x-axis.

y_str

Character string. Name of variable in data to plot on y-axis.

color_str

Character string. Name of variable in data to use as color aesthetic in plot.

fill_str

Character string. Name of variable in data to use as fill aesthetic in plot.

shape_str

Character string. Name of variable in data to use as shape aesthetic in plot.

linetype_str

Character string. Name of variable in data to use as linetype aesthetic in plot.

group_str

Character string. Name of variable in data to use as group aesthetic in plot.

fill

Fill color. Only used if fill_str is NULL.

alpha

Alpha value for transparency.